refactor: relocate services to lib directory
This commit moves the services from the root directory to a new 'lib' directory. This includes the AgendaService, PrismaService, ATSUService, and ATSUMessageService. This move is to enhance the project's structure and maintain.
