+++ src/app.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { AgendaService } from './agenda/agenda.service';
-import { PrismaService } from './prisma/prisma.service';
-import { ATSUService } from './prisma/atsu.service';
-import { ATSUMessageService } from './prisma/atsuMessage.service';
+import { AgendaService } from './lib/agenda/agenda.service';
+import { PrismaService } from './lib/prisma/prisma.service';
+import { ATSUService } from './lib/prisma/atsu.service';
+import { ATSUMessageService } from './lib/prisma/atsuMessage.service';
 import { MongooseModule } from '@nestjs/mongoose';
 
 import { DB_URL } from './config';
