+++ src/lib/agenda/definitions/atsuMessage.definitions.ts
@@ -1,4 +1,4 @@
-import { ATSUMessageService } from 'src/prisma/atsuMessage.service';
+import { ATSUMessageService } from 'src/lib/prisma/atsuMessage.service';
 
 export const defineExpireATSUMessage = (agenda, atsuMessageService: ATSUMessageService) => {
     agenda.define("expire ATSUMessage", async job => {
