+++ src/lib/types/sector.type.ts
