--- src/schemas/message.schema.ts
@@ -1,18 +1,0 @@
-import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
-import { HydratedDocument } from 'mongoose';
-
-export type MessageDocument = HydratedDocument<Message>;
-
-@Schema()
-export class Message {
-  @Prop()
-  name: string;
-
-  @Prop()
-  age: number;
-
-  @Prop()
-  breed: string;
-}
-
-export const MessageSchema = SchemaFactory.createForClass(Message);
